Переносит функции для рендера
