+++ src/main.js
@@ -9,7 +9,7 @@ import SortComponent from "./components/sort.js";
 import NoTasksComponent from "./components/no-tasks.js";
 import {generateTasks} from "./mock/task.js";
 import {generateFilters} from "./mock/filter.js";
-import {render, RenderPosition} from "./utils.js";
+import {render, RenderPosition} from "./utils/render.js";
 
 
 const TASK_COUNT = 22;
